refactor: rename file and function name
