+++ ts/features/payments/checkout/store/reducers/index.ts
@@ -17,7 +17,7 @@ import { NetworkError } from "../../../../../utils/errors";
 import { getSortedPspList } from "../../../common/utils";
 import { WalletPaymentStepEnum } from "../../types";
 import { FaultCodeCategoryEnum } from "../../types/PaymentGenericErrorAfterUserCancellationProblemJson";
-import { FaultCodeCategoryEnum as PaymentMethodNotAvailableEnum } from "../../types/PaymentMethodNotAvailableProblemJson";
+import { FaultCodeCategoryEnum as PaymentMethodNotAvailableEnum } from "../../types/PspPaymentMethodNotAvailableProblemJson";
 import { WalletPaymentFailure } from "../../types/WalletPaymentFailure";
 import {
   paymentsCalculatePaymentFeesAction,
