+++ ts/features/payments/checkout/types/WalletPaymentFailure.ts
@@ -10,7 +10,7 @@ import { ValidationFaultPaymentUnavailableProblemJson } from "../../../../../def
 import { ValidationFaultPaymentUnknownProblemJson } from "../../../../../definitions/pagopa/ecommerce/ValidationFaultPaymentUnknownProblemJson";
 import { PaymentGenericErrorAfterUserCancellationProblemJson } from "./PaymentGenericErrorAfterUserCancellationProblemJson";
 import { PaymentVerifyGenericErrorProblemJson } from "./PaymentVerifyGenericErrorProblemJson";
-import { PaymentMethodNotAvailableProblemJson } from "./PaymentMethodNotAvailableProblemJson";
+import { PspPaymentMethodNotAvailableProblemJson } from "./PspPaymentMethodNotAvailableProblemJson";
 
 export type WalletPaymentFailure = t.TypeOf<typeof WalletPaymentFailure>;
 export const WalletPaymentFailure = t.union([
@@ -25,5 +25,5 @@ export const WalletPaymentFailure = t.union([
   PaymentDuplicatedStatusFaultPaymentProblemJson,
   PaymentGenericErrorAfterUserCancellationProblemJson,
   PaymentVerifyGenericErrorProblemJson,
-  PaymentMethodNotAvailableProblemJson
+  PspPaymentMethodNotAvailableProblemJson
 ]);
